move webpack logic out of app-index, make main entrypoint for app
